requirements.txt
anyio==3.6.2
async-timeout==4.0.2
click==8.1.3
cryptography==41.0.1
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.6
pydantic==1.10.8
PyJWT==2.8.0
python-multipart==0.0.6
python-dotenv==1.0.0
redis==4.5.5
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
typing_extensions==4.6.2
#uvicorn==0.22.0
#watchfiles
uvicorn[standard]